requirements.txt
Django
South
django-decouple
django-ckeditor
django-suit
django-grappelli
django-admin-tools